refactor: move extracted resolve() helpers to recursive_helpers.go
recursive.go: 670 → 550 lines (below split threshold in CLAUDE.md)
New file: recursive_helpers.go with 5 helper functions
